Use mongo definition store for dev env set up
Use mongo and definitions-paged collection for definitions store in
dev environment set up.  This is because the trimmed mongo definition
store (mongoTrimmed, which uses definitions-trimmed collection) works
well as part of a definition store collection, but not on its own. In
comparison, definitions-paged collection stores the definitions in
entirety, and can be used as a standalone definition store.

The population of the definitions-trimmed collection is kept in place
to make this distinction between dev environment and production
deployment more evident.  To emulate the production environment more
closely, one can use dispatch+file+mongoTrimmed as
DEFINITION_STORE_PROVIDER